Oops

+++ kt/aoc2023-lib/src/commonMain/kotlin/com/github/ephemient/aoc2023/Day9.kt
@@ -5,14 +5,14 @@ class Day9(input: String) {
         line.split(' ').mapNotNull { it.toIntOrNull() }.ifEmpty { null }
     }
 
-    fun part1(): Int = nums.sumOf { it.extrapolate() }
+    fun part1(): Long = nums.sumOf { it.extrapolate() }
 
-    fun part2(): Int = nums.sumOf { it.asReversed().extrapolate() }
+    fun part2(): Long = nums.sumOf { it.asReversed().extrapolate() }
 
     companion object {
-        private fun List<Int>.extrapolate(): Int {
-            var c = 1
-            var s = 0
+        private fun List<Int>.extrapolate(): Long {
+            var c = 1L
+            var s = 0L
             for ((i, x) in withIndex()) {
                 s = c * x - s
                 c = c * (size - i) / (i + 1)
